fix: Bug in Download page
